Got a Simple server Running
The server uses the local server IP and a random port. The client and the server can connect to each other and will display a message. Runs on python 3 and was built on a Linux machine running Ubuntu. Also deleted the test files to clean up the project. You can add them back if you want to but keep them in a folder by themselves.